# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Brief Lab API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface BriefSection {
  heading: string;
  bullets: string[];
}

export interface BriefScore {
  logic?: number;
  evidence?: number;
  clarity?: number;
  originality?: number;
}

export type BriefResultQuality =
  (typeof BriefResultQuality)[keyof typeof BriefResultQuality];

export const BriefResultQuality = {
  good: "good",
  filler: "filler",
} as const;

export interface BriefResult {
  quality: BriefResultQuality;
  sections: BriefSection[];
  score?: BriefScore;
  summary?: string;
  message?: string;
}

export type BriefGenerateInputBriefType =
  (typeof BriefGenerateInputBriefType)[keyof typeof BriefGenerateInputBriefType];

export const BriefGenerateInputBriefType = {
  College_Essay_Outline: "College Essay Outline",
  Policy_Memo: "Policy Memo",
  Debate_Case: "Debate Case",
} as const;

export type BriefGenerateInputTone =
  (typeof BriefGenerateInputTone)[keyof typeof BriefGenerateInputTone];

export const BriefGenerateInputTone = {
  Neutral_Analytic: "Neutral Analytic",
  Advocacy: "Advocacy",
  Academic_Formal: "Academic Formal",
} as const;

export type BriefGenerateInputMode =
  (typeof BriefGenerateInputMode)[keyof typeof BriefGenerateInputMode];

export const BriefGenerateInputMode = {
  generate: "generate",
  polish: "polish",
} as const;

export interface BriefGenerateInput {
  /** @minLength 1 */
  prompt: string;
  sourceNotes?: string;
  briefType: BriefGenerateInputBriefType;
  tone: BriefGenerateInputTone;
  mode?: BriefGenerateInputMode;
}

export interface Brief {
  id: number;
  firebaseUid: string;
  title: string;
  briefType: string;
  tone: string;
  outlineText: string;
  sections?: string;
  /** @nullable */
  score?: string | null;
  summary?: string;
  tags?: string;
  prompt?: string;
  sourceNotes?: string;
  createdAt: string;
  updatedAt?: string;
}

export interface BriefInput {
  firebaseUid: string;
  title: string;
  briefType: string;
  tone: string;
  outlineText: string;
  sections?: string;
  score?: string;
  summary?: string;
  tags?: string;
  prompt: string;
  sourceNotes?: string;
}

export interface BriefUpdate {
  title?: string;
  briefType?: string;
  tone?: string;
  outlineText?: string;
  sections?: string;
  score?: string;
  summary?: string;
  tags?: string;
}

export type BriefStatsByType = { [key: string]: number };

export interface BriefStats {
  total: number;
  byType: BriefStatsByType;
  recentCount: number;
  topTags?: string[];
}

export type ListBriefsParams = {
  firebaseUid: string;
  briefType?: string;
  search?: string;
};

export type GetBriefStatsParams = {
  firebaseUid: string;
};