# lib/api-zod/src/generated/types/briefGenerateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Brief Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { BriefGenerateInputBriefType } from "./briefGenerateInputBriefType";
import type { BriefGenerateInputMode } from "./briefGenerateInputMode";
import type { BriefGenerateInputTone } from "./briefGenerateInputTone";

export interface BriefGenerateInput {
  /** @minLength 1 */
  prompt: string;
  sourceNotes?: string;
  briefType: BriefGenerateInputBriefType;
  tone: BriefGenerateInputTone;
  mode?: BriefGenerateInputMode;
}