# lib/api-zod/src/generated/types/briefInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Brief Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface BriefInput {
  firebaseUid: string;
  title: string;
  briefType: string;
  tone: string;
  outlineText: string;
  sections?: string;
  score?: string;
  summary?: string;
  tags?: string;
  prompt: string;
  sourceNotes?: string;
}