# lib/api-zod/src/generated/types/briefResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Brief Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { BriefResultQuality } from "./briefResultQuality";
import type { BriefScore } from "./briefScore";
import type { BriefSection } from "./briefSection";

export interface BriefResult {
  quality: BriefResultQuality;
  sections: BriefSection[];
  score?: BriefScore;
  summary?: string;
  message?: string;
}