# lib/api-zod/src/generated/types/briefScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Brief Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface BriefScore {
  logic?: number;
  evidence?: number;
  clarity?: number;
  originality?: number;
}