# lib/api-zod/src/generated/types/briefStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * College Brief Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { BriefStatsByType } from "./briefStatsByType";

export interface BriefStats {
  total: number;
  byType: BriefStatsByType;
  recentCount: number;
  topTags?: string[];
}